Add files via upload
"This repository contains a Library Management System project that is open-source and free to use. Anyone is welcome to contribute and make modifications to the code as needed. The aim of this project is to provide an efficient way of managing a library's books and resources. Feel free to use and customize the code to fit your specific library management needs